agents-mcp: fold credentials/types into auth-store
CI's tsdown DTS bundle step couldn't resolve `src/credentials/types.d.ts`
under parallel-workspace build conditions, even though the file
existed and the local single-package build succeeded. Moving the
shared `AuthStore` / `AuthStoreHooks` types alongside their only
implementation (`auth-store.ts`) sidesteps the cross-file DTS
resolution and keeps the surface tighter — only one importer
(`auth/sdk-provider`) needed updating.
