+++ packages/agents-mcp/src/credentials/auth-store.ts
@@ -1,8 +1,27 @@
 // Internal — do not export from src/index.ts. Instantiated once per
 // registry; lives in-process. Cross-process persistence is the operator's
-// concern via the hooks declared on the auth config (see types.ts).
+// concern via the hooks declared on the auth config.
 import type { OAuthClientInfo, OAuthTokens } from '../types'
-import type { AuthStore, AuthStoreHooks } from './types'
+
+/**
+ * Internal token + client cache used by the registry. Created per-registry,
+ * never crossed by the public API. Seeded from `auth.tokens` / `auth.client`
+ * on `addServer`; mutated as the SDK refreshes / completes DCR. Calls into
+ * the per-server `onTokensChanged` / `onClientRegistered` callbacks declared
+ * on the auth config so the operator can persist if they want.
+ */
+export interface AuthStore {
+  getOAuthTokens(server: string): OAuthTokens | undefined
+  saveOAuthTokens(server: string, tokens: OAuthTokens): Promise<void>
+  getOAuthClientInfo(server: string): OAuthClientInfo | undefined
+  saveOAuthClientInfo(server: string, info: OAuthClientInfo): Promise<void>
+}
+
+/** Per-server hooks registered on `addServer` and invoked on cache mutations. */
+export interface AuthStoreHooks {
+  onTokensChanged?: (tokens: OAuthTokens) => void | Promise<void>
+  onClientRegistered?: (client: OAuthClientInfo) => void | Promise<void>
+}
 
 export interface InternalAuthStore extends AuthStore {
   /** Pre-seed the cache with tokens read from `auth.tokens`. */
